# Improve interpolate() speed for channels_last CPU images and masks