@fmassa 2022-10-06
@datumbox there is further opportunities for optimization for `uint8_t` types and 3 / 1 channels by leveraging explicit vectorization, but this would involve some significant work