@NicolasHug 2022-10-06
Thanks all for the reviews!

> I would maybe prefer having a helper

done!

> Did you see similar opportunities for improvement on the downsample kernels?

Yes, this PR addresses upsampling and downsampling (the name of the kernels is a bit misleading!)

> Also how about bicubic interpolation?

Echoing comments from Victor and Francisco above: the trick we're using here isn't directly applicable to bicubic interpolation. It can be applied to the linear, nearest and nearest-exact kernels for 3D (temporal) inputs though, which is something I'll look into as well once this PR is merged.

There are various opportunities for improvement over the current implementations, but most of them would require a substantial amount of work.