@NicolasHug 2022-10-11
I'm on it @datumbox . Failure is due to a test in pytorchvideo which makes a hard `assert ==` on float values instead of `assert_close()`.

I don't know whether this PR should have been reverted manually though, considering the internal revert diff still hasn't landed D40255846