@datumbox 2022-10-11
Thanks for looking into it @NicolasHug. FYI there are some additional [tests](https://www.internalfb.com/intern/testinfra/diagnostics/844425196841106.844425004488050.1665471605/) failing unrelated to PyTorch Video:
```
terminate called after throwing an instance of 'c10::Error'
  what():  dtype 'Float' not selected for kernel tag upsample_nearest2d_channels_last
```