@NicolasHug 2022-10-11
These errors are also present on the revert diff, so I'm not sure it's immediately relevant to the changes