@malfet 2022-10-11
Hmm, I'm a bit confused about what happened: as I can see code landed internally as [D40217870](https://www.internalfb.com/diff/D40217870) but reverted in OSS.